+++ spec/spec_helper.rb
@@ -1,3 +1,9 @@
+require 'simplecov'
+SimpleCov.start do
+  add_filter '/spec/'
+  minimum_coverage 85
+end
+
 require 'active_postgrest'
 
 RSpec.configure do |config|